refactor: home画面をpagesスライスに移動
