+++ src/pages/home/types.ts
