+++ src/pages/home/ui/background-preview-lobby.tsx
