+++ src/pages/home/ui/study-button.tsx
